Refactor Counter entry to be with CounterEntry
