+++ server/src/Command/GetCommand.cpp
@@ -28,16 +28,12 @@ namespace Sider::Command
                 return Result::nil();
             }
 
-            if (entry->type() == Entry::Type::KEEPER || entry->type() == Entry::Type::QUEUE) {
-                return Result::with(entry->get());
-            }
-
             switch (entry->type()) {
-                case Entry::Type::COUNTER: {
-                    Entry::CounterEntry* counterEntry = dynamic_cast<Entry::CounterEntry*>(entry);
+                case Entry::Type::COUNTER:
+                case Entry::Type::KEEPER:
+                case Entry::Type::QUEUE:
+                    return Result::with(entry->get());
 
-                    return Result::with(counterEntry->get());
-                }
                 case Entry::Type::RATER: {
                     Entry::RaterEntry* raterEntry = dynamic_cast<Entry::RaterEntry*>(entry);
